Feat/#170 [FIX] testTime을 Long으로 통일

+++ src/main/java/swm_nm/morandi/test/dto/TestRecordDto.java
@@ -16,7 +16,7 @@ import java.util.Map;
 @ToString
 public class TestRecordDto {
     private LocalDateTime testDate;
-    private Integer testTime;
+    private Long testTime;
     private Integer problemCount;
     private DifficultyLevel startDifficulty;
     private DifficultyLevel endDifficulty;
